campos numeric em tabelas

+++ models/autorizacaoInternacaoHospitalarModels.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, Float, Date, BigInteger
+from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, Numeric, Date, BigInteger
 from sqlalchemy.orm import relationship
 from conexao.conect_db import Base
 
@@ -27,7 +27,7 @@ class AutorizacaoInternacaoHospitalar(Base):
     ProcedimentoSolicitado = Column(BigInteger, nullable=True)
     CaraterInternacao = Column(Integer, ForeignKey("aux_carater_internacao.id"), nullable=True)
     carater_internacao = relationship("CaraterInternacao", back_populates="autorizacao_internacao")
-    MotivoSaida = Column(Integer, ForeignKey("aux_motivo_saida.id"), nullable=True)
+    MotivoSaida = Column(Numeric(1,1), ForeignKey("aux_motivo_saida.id"), nullable=True)
     motivo_saida = relationship("MotivoSaida", back_populates="autorizacao_internacao")
     CNSSolicitante = Column(BigInteger, nullable=True)
     CNSResponsavel = Column(BigInteger, nullable=True)
